+++ src/main/java/com/verdantartifice/thaumicwonders/ThaumicWonders.java
@@ -21,7 +21,7 @@ import org.apache.logging.log4j.Logger;
 public class ThaumicWonders {
     public static final String MODID = "thaumicwonders";
     public static final String NAME = "Thaumic Wonders Unofficial";
-    public static final String VERSION = "2.1.1";
+    public static final String VERSION = "2.2.0";
     public static final String DEPENDENCIES = "required-after:thaumcraft";
 
     public static final CreativeTabs CREATIVE_TAB = new CreativeTabTW(CreativeTabs.getNextID(), ThaumicWonders.MODID);
@@ -57,7 +57,12 @@ public class ThaumicWonders {
 
     //TODO: Future Release
     //  Thauma Llama
-    //  Check out Gadomancy to see if any additions can be made to TW (specifically the Infusion Claw for infusion automation)
+    //  Shearing golem seal
+    //  Fishing golem seal
+    //  Crafting golem seal
+    //  Eldritch cluster harvest infusion enchantment (or just a single pickaxe that does it
+    //  Register for eldritch cluster harvesting
+    //  Item retriveal golem seal (like corporea crystal)
     //  Heximite explosion converts blocks into tainted variants (add CrT/GS integration for additional conversions)
     //  Research Brews + Crucible addon - create brews that randomly grant research after being consumed.
     //  Teleport Home focus effect