+++ src/main/java/com/verdantartifice/thaumicwonders/common/utils/NBTHelper.java
@@ -11,19 +11,4 @@ public class NBTHelper {
             stack.setTagCompound(new NBTTagCompound());
         }
     }
-
-    public static NBTTagCompound serializeBlockPos(BlockPos pos) {
-        NBTTagCompound tag = new NBTTagCompound();
-        tag.setInteger("posX", pos.getX());
-        tag.setInteger("posY", pos.getY());
-        tag.setInteger("posZ", pos.getZ());
-        return tag;
-    }
-
-    public static BlockPos deserializeBlockPos(NBTTagCompound tag) {
-        int x = tag.getInteger("posX");
-        int y = tag.getInteger("posY");
-        int z = tag.getInteger("posZ");
-        return new BlockPos(x, y, z);
-    }
 }