+++ src/main/java/com/verdantartifice/thaumicwonders/common/utils/StringHelper.java
@@ -3,6 +3,8 @@ package com.verdantartifice.thaumicwonders.common.utils;
 import com.verdantartifice.thaumicwonders.ThaumicWonders;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.util.text.TextComponentTranslation;
+import net.minecraft.world.DimensionType;
+import net.minecraftforge.common.DimensionManager;
 
 public class StringHelper {
 
@@ -20,5 +22,21 @@ public class StringHelper {
 
     public static TextComponentTranslation getTranslatedTextComponent(String unlocName, String type, String... params) {
         return new TextComponentTranslation(getTranslationKey(unlocName, type, params));
+    }
+
+    public static String getDimensionName(int dimensionId) {
+        if (!DimensionManager.isDimensionRegistered(dimensionId)) {
+            return Integer.toString(dimensionId);
+        }
+        DimensionType type = DimensionManager.getProviderType(dimensionId);
+        if (type == null) {
+            return Integer.toString(dimensionId);
+        }
+        String name = type.getName();
+        int[] dims = DimensionManager.getDimensions(type);
+        if (dims != null && dims.length > 1) {
+            name += " " + dimensionId;
+        }
+        return name;
     }
 }